requirements.txt
# Phase 2 Requirements
openai>=1.0.0
python-dotenv>=1.0.0

# Phase 3 Requirements
chromadb>=0.4.0
